Remove Compaction strategy
