+++ src/datanode/src/instance.rs
@@ -44,7 +44,7 @@ use query::query_engine::{QueryEngineFactory, QueryEngineRef};
 use servers::Mode;
 use session::context::QueryContext;
 use snafu::prelude::*;
-use storage::compaction::{CompactionHandler, CompactionSchedulerRef, LeveledPicker};
+use storage::compaction::{CompactionHandler, CompactionSchedulerRef, LeveledTimeWindowPicker};
 use storage::config::EngineConfig as StorageEngineConfig;
 use storage::scheduler::{LocalScheduler, SchedulerConfig};
 use storage::EngineImpl;
@@ -395,7 +395,7 @@ impl Instance {
 }
 
 fn create_compaction_scheduler<S: LogStore>(opts: &DatanodeOptions) -> CompactionSchedulerRef<S> {
-    let picker = LeveledPicker::default();
+    let picker = LeveledTimeWindowPicker::default();
     let config = SchedulerConfig::from(opts);
     let handler = CompactionHandler { picker };
     let scheduler = LocalScheduler::new(config, handler);
