+++ src/storage/src/compaction/writer.rs
@@ -29,8 +29,7 @@ pub(crate) async fn build_sst_reader(
     schema: RegionSchemaRef,
     sst_layer: AccessLayerRef,
     files: &[FileHandle],
-    lower_sec_inclusive: i64,
-    upper_sec_exclusive: i64,
+    time_range: (Option<i64>, Option<i64>),
 ) -> error::Result<ChunkReaderImpl> {
     // TODO(hl): Schemas in different SSTs may differ, thus we should infer
     // timestamp column name from Parquet metadata.
@@ -43,14 +42,9 @@ pub(crate) async fn build_sst_reader(
     ChunkReaderBuilder::new(region_id, schema, sst_layer)
         .pick_ssts(files)
         .filters(
-            build_time_range_filter(
-                lower_sec_inclusive,
-                upper_sec_exclusive,
-                &ts_col_name,
-                ts_col_unit,
-            )
-            .into_iter()
-            .collect(),
+            build_time_range_filter(time_range, &ts_col_name, ts_col_unit)
+                .into_iter()
+                .collect(),
         )
         .build()
         .await
@@ -59,21 +53,22 @@ pub(crate) async fn build_sst_reader(
 /// Build time range filter expr from lower (inclusive) and upper bound(exclusive).
 /// Returns `None` if time range overflows.
 fn build_time_range_filter(
-    low_sec: i64,
-    high_sec: i64,
+    time_range: (Option<i64>, Option<i64>),
     ts_col_name: &str,
     ts_col_unit: TimeUnit,
 ) -> Option<Expr> {
-    debug_assert!(low_sec <= high_sec);
+    let (low_ts_inclusive, high_ts_exclusive) = time_range;
     let ts_col = DfExpr::Column(datafusion_common::Column::from_name(ts_col_name));
 
     // Converting seconds to whatever unit won't lose precision.
     // Here only handles overflow.
-    let low_ts = common_time::Timestamp::new_second(low_sec)
-        .convert_to(ts_col_unit)
+    let low_ts = low_ts_inclusive
+        .map(common_time::Timestamp::new_second)
+        .and_then(|ts| ts.convert_to(ts_col_unit))
         .map(|ts| ts.value());
-    let high_ts = common_time::Timestamp::new_second(high_sec)
-        .convert_to(ts_col_unit)
+    let high_ts = high_ts_exclusive
+        .map(common_time::Timestamp::new_second)
+        .and_then(|ts| ts.convert_to(ts_col_unit))
         .map(|ts| ts.value());
 
     let expr = match (low_ts, high_ts) {
@@ -296,8 +291,7 @@ mod tests {
             schema,
             sst_layer,
             files,
-            lower_sec_inclusive,
-            upper_sec_exclusive,
+            (Some(lower_sec_inclusive), Some(upper_sec_exclusive)),
         )
         .await
         .unwrap();
@@ -378,9 +372,15 @@ mod tests {
         sst_layer: AccessLayerRef,
     ) -> Vec<i64> {
         let mut timestamps = vec![];
-        let mut reader = build_sst_reader(REGION_ID, schema, sst_layer, files, i64::MIN, i64::MAX)
-            .await
-            .unwrap();
+        let mut reader = build_sst_reader(
+            REGION_ID,
+            schema,
+            sst_layer,
+            files,
+            (Some(i64::MIN), Some(i64::MAX)),
+        )
+        .await
+        .unwrap();
         while let Some(chunk) = reader.next_chunk().await.unwrap() {
             let ts = chunk.columns[0]
                 .as_any()
@@ -447,8 +447,7 @@ mod tests {
             schema.clone(),
             sst_layer.clone(),
             &input_files,
-            0,
-            3,
+            (Some(0), Some(3)),
         )
         .await
         .unwrap();
@@ -457,8 +456,7 @@ mod tests {
             schema.clone(),
             sst_layer.clone(),
             &input_files,
-            3,
-            6,
+            (Some(3), Some(6)),
         )
         .await
         .unwrap();
@@ -467,8 +465,7 @@ mod tests {
             schema.clone(),
             sst_layer.clone(),
             &input_files,
-            6,
-            10,
+            (Some(6), Some(10)),
         )
         .await
         .unwrap();
@@ -554,7 +551,12 @@ mod tests {
 
     #[test]
     fn test_build_time_range_filter() {
-        assert!(build_time_range_filter(i64::MIN, i64::MAX, "ts", TimeUnit::Nanosecond).is_none());
+        assert!(build_time_range_filter(
+            (Some(i64::MIN), Some(i64::MAX)),
+            "ts",
+            TimeUnit::Nanosecond
+        )
+        .is_none());
 
         assert_eq!(
             Expr::from(datafusion_expr::binary_expr(
@@ -562,10 +564,10 @@ mod tests {
                 Operator::Lt,
                 datafusion_expr::lit(timestamp_to_scalar_value(
                     TimeUnit::Nanosecond,
-                    Some(TimeUnit::Second.factor() as i64 / TimeUnit::Nanosecond.factor() as i64)
-                ))
+                    Some(TimeUnit::Second.factor() as i64 / TimeUnit::Nanosecond.factor() as i64),
+                )),
             )),
-            build_time_range_filter(i64::MIN, 1, "ts", TimeUnit::Nanosecond).unwrap()
+            build_time_range_filter((Some(i64::MIN), Some(1)), "ts", TimeUnit::Nanosecond).unwrap()
         );
 
         assert_eq!(
@@ -576,10 +578,10 @@ mod tests {
                     TimeUnit::Nanosecond,
                     Some(
                         2 * TimeUnit::Second.factor() as i64 / TimeUnit::Nanosecond.factor() as i64
-                    )
-                ))
+                    ),
+                )),
             )),
-            build_time_range_filter(2, i64::MAX, "ts", TimeUnit::Nanosecond).unwrap()
+            build_time_range_filter((Some(2), Some(i64::MAX)), "ts", TimeUnit::Nanosecond).unwrap()
         );
     }
 }
